강의계획서 완료 보고서 마무리중

+++ src/main/java/kr/co/ChimAcademy/entity/MemberEntity.java
@@ -42,7 +42,7 @@ public class MemberEntity implements Serializable{
 	private int sYear;
 	private int gender;
 	private int totCredit;
-	private int avgCredit;
+	private double avgCredit;
 	private String hp;
 	private String email;
 	private int status;
